Removed user password setting
There is a limited number of attempts for the unattended password set for a user with `dscl ` command. Once they depleted, `dscl` command requires entering the previous user password explicitly.
The forced setting of a random password with `dscl ` command on the first launch does not make much sense because it's still impossible to ScreenShare to the system when there is no password. But that action (force password set) just subtracts the password reset attempt from the total number of possible unattended resets, which makes troubleshooting and custom provisioning a bit more difficult.