feat: add command to grow root APFS volume
The current EC2 macOS AMIs default to an APFS volume size of 100 GB.
Volumes with larger sizes are not automatically usable when the instance
becomes available.

This change adds a new command to use ec2-macos-utils to grow the root
APFS volume to its maximum size. This will allow 100% of the volumes
size to be available when the instance becomes available.