fix: detect module even when false is set
This was tripping up on false being a zero value. When comparing (with
cmp) against the zero value of the struct, the _desired_ value is
confused with the zero value.

Now the module uses a pointer to detect whether or not a value has been
set to handle the case that `false` is the intended setting.

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>